# contracts/bpi-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FT MES BPI API",
    "version": "1.0.0-shadow",
    "description": "智能批次与工艺数据中心同步 API。Phase 1 仅允许影子批次，不触发 WOM/QCS/WMS 写操作。"
  },
  "servers": [
    {"url": "/", "description": "MES gateway or local simulator"}
  ],
  "security": [{"bearerAuth": []}],
...
  "paths": {
...
    "/bpi/v1/point-catalog/snapshots": {
      "get": {
        "tags": ["Point Catalog"], "operationId": "listPointCatalogSnapshots", "summary": "查询不可变点位目录快照",
        "parameters": [{"$ref": "#/components/parameters/PlantId"}, {"$ref": "#/components/parameters/LineIdQuery"}],
        "responses": {"200": {"description": "快照列表", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PointCatalogSnapshotListResponse"}}}}}
      },
      "post": {
        "tags": ["Point Catalog"], "operationId": "importPointCatalogSnapshot", "summary": "导入外部点位目录快照",
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}, {"$ref": "#/components/parameters/IfMatch"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PointCatalogSnapshotCommand"}}}},
        "responses": {
          "200": {"description": "已导入快照", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PointCatalogResponse"}}}},
          "409": {"$ref": "#/components/responses/Conflict"},
          "428": {"$ref": "#/components/responses/PreconditionRequired"}
        }
      }
    },
    "/bpi/v1/point-catalog/current": {
      "get": {
        "tags": ["Point Catalog"], "operationId": "getCurrentPointCatalog", "summary": "查询产线当前点位目录和准入状态",
        "parameters": [
          {"name": "plantId", "in": "query", "required": true, "schema": {"type": "string"}},
          {"name": "lineId", "in": "query", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {"200": {"description": "当前快照；未导入时 data 为 null", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PointCatalogCurrentResponse"}}}}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "parameters": {
      "PlantId": {"name": "plantId", "in": "query", "required": true, "schema": {"type": "string"}},
...
      "LineIdQuery": {"name": "lineId", "in": "query", "schema": {"type": "string"}},
...
      "IdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 8, "maxLength": 128}},
      "IfMatch": {"name": "If-Match", "in": "header", "required": true, "description": "Current numeric entity revision", "schema": {"type": "string", "pattern": "^(W/)?\\\"?[0-9]+\\\"?$"}}
    },
...
    "responses": {
...
      "Conflict": {"description": "revision 冲突或对象已处理", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}},
      "PreconditionRequired": {"description": "缺少 Idempotency-Key 或 If-Match", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}},
...
    },
    "schemas": {
      "ResponseMeta": {
        "type": "object", "required": ["traceId", "generatedAt"],
        "properties": {"traceId": {"type": "string"}, "generatedAt": {"type": "string", "format": "date-time"}, "snapshotAt": {"type": ["string", "null"], "format": "date-time"}, "nextCursor": {"type": ["string", "null"]}}
      },
      "Problem": {
        "type": "object", "required": ["type", "title", "status", "detail", "traceId"],
        "properties": {"type": {"type": "string", "format": "uri-reference"}, "title": {"type": "string"}, "status": {"type": "integer"}, "detail": {"type": "string"}, "traceId": {"type": "string"}, "currentRevision": {"type": ["integer", "null"]}, "errors": {"type": "array", "items": {"type": "object"}}}
      },
...
      "PointCatalogPointCommand": {"type": "object", "required": ["productId", "deviceId", "propertyId", "deviceState", "registered", "propertyPresent", "calibrationStatus", "sourceSequenceEnabled"], "properties": {"localityGroup": {"type": ["string", "null"], "maxLength": 128}, "productId": {"type": "string", "maxLength": 128}, "deviceId": {"type": "string", "maxLength": 128}, "propertyId": {"type": "string", "maxLength": 128, "description": "Canonical property identity emitted by the exporter and consumed by BPI."}, "sourcePropertyId": {"type": ["string", "null"], "maxLength": 128, "description": "Original JetLinks/source property identity before exporter normalization."}, "pointName": {"type": ["string", "null"], "maxLength": 256}, "unit": {"type": ["string", "null"], "maxLength": 32}, "dataType": {"type": ["string", "null"], "maxLength": 64}, "deviceState": {"type": "string", "enum": ["ACTIVE", "INACTIVE", "UNKNOWN"]}, "registered": {"type": "boolean"}, "propertyPresent": {"type": "boolean"}, "calibrationVersion": {"type": ["string", "null"], "maxLength": 128}, "calibrationStatus": {"type": "string", "enum": ["VERIFIED", "UNVERIFIED", "MISSING"]}, "sourceSequenceEnabled": {"type": "boolean"}}},
      "PointCatalogSnapshotCommand": {"type": "object", "required": ["source", "sourceInstance", "sourceRevision", "plantId", "lineId", "observedAt", "points", "reason"], "properties": {"source": {"type": "string", "maxLength": 64, "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$"}, "sourceInstance": {"type": "string", "maxLength": 128}, "sourceRevision": {"type": "string", "maxLength": 128}, "plantId": {"type": "string", "maxLength": 64}, "lineId": {"type": "string", "maxLength": 128}, "observedAt": {"type": "string", "format": "date-time"}, "points": {"type": "array", "maxItems": 10000, "items": {"$ref": "#/components/schemas/PointCatalogPointCommand"}}, "reason": {"type": "string", "minLength": 3, "maxLength": 500}}},
      "PointCatalogSnapshot": {"type": "object", "required": ["id", "source", "sourceInstance", "sourceRevision", "plantId", "lineId", "checksum", "observedAt", "pointCount", "readyPointCount", "importedBy", "importedAt"], "properties": {"id": {"type": "string", "format": "uuid"}, "source": {"type": "string"}, "sourceInstance": {"type": "string"}, "sourceRevision": {"type": "string"}, "plantId": {"type": "string"}, "lineId": {"type": "string"}, "checksum": {"type": "string", "pattern": "^[a-f0-9]{64}$"}, "observedAt": {"type": "string", "format": "date-time"}, "pointCount": {"type": "integer", "minimum": 0}, "readyPointCount": {"type": "integer", "minimum": 0}, "importedBy": {"type": "string"}, "importedAt": {"type": "string", "format": "date-time"}}},
      "PointCatalogPoint": {"allOf": [{"$ref": "#/components/schemas/PointCatalogPointCommand"}, {"type": "object", "required": ["id", "snapshotId", "plantId", "lineId", "ready", "readinessIssues"], "properties": {"id": {"type": "string", "format": "uuid"}, "snapshotId": {"type": "string", "format": "uuid"}, "plantId": {"type": "string"}, "lineId": {"type": "string"}, "ready": {"type": "boolean"}, "readinessIssues": {"type": "array", "items": {"type": "string"}}}}]},
      "PointCatalog": {"type": "object", "required": ["snapshot", "points"], "properties": {"snapshot": {"$ref": "#/components/schemas/PointCatalogSnapshot"}, "points": {"type": "array", "items": {"$ref": "#/components/schemas/PointCatalogPoint"}}}},
      "PointCatalogResponse": {"type": "object", "required": ["data", "meta"], "properties": {"data": {"$ref": "#/components/schemas/PointCatalog"}, "meta": {"$ref": "#/components/schemas/ResponseMeta"}}},
      "PointCatalogCurrentResponse": {"type": "object", "required": ["data", "meta"], "properties": {"data": {"oneOf": [{"$ref": "#/components/schemas/PointCatalog"}, {"type": "null"}]}, "meta": {"$ref": "#/components/schemas/ResponseMeta"}}},
      "PointCatalogSnapshotListResponse": {"type": "object", "required": ["data", "meta"], "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/PointCatalogSnapshot"}}, "meta": {"$ref": "#/components/schemas/ResponseMeta"}}},
...
    }
  }
}